Update. Fixed auctioned items not to have random enchantments generated when they are expected to be there. Guaranteed disenchanting spoils prediction. Made it possible to find epic crafted items but very rarely.

+++ Assets/Resources/Source/Game/World/Auctionable.cs
@@ -12,7 +12,7 @@ public class Auctionable
     public int maxPrice;
 
     //Chance to see this item enter market each hour
-    public int frequency;
+    public float frequency;
 
     //Currently opened auctionable
     public static Auctionable auctionable;
